joycaption-mcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "joycaption-mcp"
version = "1.0.0"
description = "MCP server for JoyCaption image captioning"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mcp>=0.9.0",
    "torch>=2.0.0",
    "torchvision",
    "transformers>=4.36.0",
    "Pillow>=10.0.0",
    "accelerate>=0.25.0",
]

[project.scripts]
joycaption-mcp = "joycaption_mcp:main"